simplify based on assumptions
